+++ src/module/api.ts
@@ -61,7 +61,7 @@ const API = {
     // TODO find a better method than this
     let lastElement = '';
     let specialCaseEndWith1 = false;
-    if(sourceToken.name.endsWith(')')){
+    if (sourceToken.name.endsWith(')')) {
       specialCaseEndWith1 = true;
     }
     const matches = <any[]>sourceToken.name.match(/(?<=\().+?(?=\))/g);
@@ -70,7 +70,7 @@ const API = {
     } else {
       lastElement = sourceToken.name;
     }
-    if(specialCaseEndWith1){
+    if (specialCaseEndWith1) {
       lastElement = lastElement + ')';
     }
 
@@ -80,7 +80,6 @@ const API = {
       }) || undefined;
 
     if (removePolymorpher) {
-
       const updatesForRevert: any = sourceToken.actor?.getFlag(
         CONSTANTS.MODULE_NAME,
         PolymorpherFlags.UPDATES_FOR_REVERT,
@@ -91,7 +90,9 @@ const API = {
         warn(`Can't revert this token without the flag '${PolymorpherFlags.UPDATES_FOR_REVERT}'`, true);
         return;
       }
-      const arrayMutationNames:string[] = <string[]>actor?.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.MUTATION_NAMES_FOR_REVERT);
+      const arrayMutationNames: string[] = <string[]>(
+        actor?.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.MUTATION_NAMES_FOR_REVERT)
+      );
       await sourceToken.actor?.unsetFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.MUTATION_NAMES_FOR_REVERT);
       await sourceToken.actor?.unsetFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.UPDATES_FOR_REVERT);
 
@@ -105,8 +106,8 @@ const API = {
       const animation = polyData?.animation;
       tokenData = updatesForRevert.tokenData || tokenData;
       tokenFromTransform = <Token>canvas.tokens?.placeables.find((t: Token) => {
-        return t.id === tokenData._id;
-      }) || tokenData;
+          return t.id === tokenData._id;
+        }) || tokenData;
 
       if (animationExternal && animationExternal.sequence) {
         //@ts-ignore
@@ -127,7 +128,7 @@ const API = {
         //@ts-ignore
         actor?.revertOriginalForm();
       } else {
-        for(const revertName of arrayMutationNames){
+        for (const revertName of arrayMutationNames) {
           info(`Revert token ${sourceToken.name}`);
           //@ts-ignore
           await warpgate.revert(sourceToken.document, revertName);