+++ src/module/settings.ts
@@ -119,4 +119,13 @@ export const registerSettings = function () {
       Bottom: 'Bottom',
     },
   });
+
+  game.settings.register(CONSTANTS.MODULE_NAME, 'hudColorButton', {
+    name: i18n(CONSTANTS.MODULE_NAME + '.settings.hudColorButton.title'),
+    hint: i18n(CONSTANTS.MODULE_NAME + '.settings.hudColorButton.hint'),
+    scope: 'client',
+    type: String,
+    default: '#b8860b',
+    config: true,
+  });
 };